+++ src/main/java/lumien/randomthings/config/Features.java
@@ -32,6 +32,9 @@ public class Features {
 			comment = "Whether the Spectre Sapling should be enabled. If disabled, the sapling will not grow and cannot be created from regular saplings.") public static boolean ENABLE_SPECTRE_SAPLING =
 					true;
 
+	@ConfigOption(category = CATEGORY, name = "AncientBrickDropItems", comment = "Whether Ancient Brick blocks can be broken and will drop their item forms.")
+	public static boolean ANCIENT_BRICK_DROP_ITEMS = false;
+
 	public static final String SUMMONING_PENDULUM_BLACKLIST_COMMENT = "List of entity resource locations that cannot be captured by the Summoning Pendulum. Format: modid:entityname. Example: minecraft:villager, randomthings:spirit.";
 
 	public static String[] SUMMONING_PENDULUM_BLACKLIST = new String[0];